requirements.txt
supabase
playwright
python-dotenv
